% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcpa.r
\name{lcpa_glpk}
\alias{lcpa_glpk}
\title{Lattice Cutting Plane Algorithm using GLPK}
\usage{
lcpa_glpk(x, y, R_max, time_limit)
}
\arguments{
\item{x}{feature matrix}

\item{y}{response vector}

\item{R_max}{max number of features (including intercept)}

\item{time_limit}{max running time in seconds}
}
\description{
Lattice Cutting Plane Algorithm using GLPK
}
