% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskslimr.glpk-package.R
\docType{package}
\name{riskslimr.glpk-package}
\alias{riskslimr.glpk}
\alias{riskslimr.glpk-package}
\title{riskslimr.glpk: GLPK backend for riskslimr}
\description{
GLPK backend for riskslimr
}
\author{
\strong{Maintainer}: Robert Zamora \email{zamora.rr@gmail.com}

}
\keyword{internal}
